# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ottiya Korean API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface SendOpenaiVoiceMessageBody {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

/**
 * Which character is speaking (determines voice)
 */
export type TtsRequestCharacter =
  (typeof TtsRequestCharacter)[keyof typeof TtsRequestCharacter];

export const TtsRequestCharacter = {
  drColi: "drColi",
  bori: "bori",
  narrator: "narrator",
} as const;

export interface TtsRequest {
  text: string;
  /** Which character is speaking (determines voice) */
  character: TtsRequestCharacter;
}

export interface TtsResponse {
  /** Base64-encoded MP3 audio */
  audio: string;
}

export interface SttRequest {
  /** Base64-encoded audio from mic */
  audio: string;
}

export interface SttResponse {
  transcript: string;
  confidence?: number;
}

export interface Episode {
  id: string;
  title: string;
  sceneCount: number;
  vocabulary: string[];
}

export type EpisodeDetailScenesItem = { [key: string]: unknown };

export interface EpisodeDetail {
  id: string;
  title: string;
  background: string;
  scenes: EpisodeDetailScenesItem[];
}

export interface ProgressEntry {
  id: number;
  childId: string;
  episodeId: string;
  sceneId: string;
  stars: number;
  completedAt: string;
}

export interface ChildProgress {
  childId: string;
  completedScenes: ProgressEntry[];
  episodesStarted: number;
  episodesCompleted: number;
  totalStars: number;
}

export interface SaveProgressBody {
  episodeId: string;
  sceneId: string;
  stars: number;
}

export interface LogbookEntry {
  id: number;
  childId: string;
  korean: string;
  english: string;
  romanization: string;
  imageUrl?: string;
  episodeId?: string;
  addedAt: string;
}

export interface AddLogbookEntryBody {
  korean: string;
  english: string;
  romanization: string;
  imageUrl?: string;
  episodeId?: string;
}

export interface LogbookStats {
  totalWords: number;
  wordsThisWeek: number;
  recentWords: LogbookEntry[];
}

export type ChatRequestCharacter =
  (typeof ChatRequestCharacter)[keyof typeof ChatRequestCharacter];

export const ChatRequestCharacter = {
  drColi: "drColi",
  bori: "bori",
} as const;

export interface ChatRequest {
  message: string;
  character: ChatRequestCharacter;
  episodeId?: string;
  childName?: string;
  childAge?: number;
  /** Structured summary of the child's lesson progress for adaptive AI responses */
  progressSummary?: string;
}

export interface ChatResponse {
  reply: string;
  koreanWord?: string;
  englishWord?: string;
  romanization?: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};