# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ottiya Korean API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestCharacter } from "./chatRequestCharacter";

export interface ChatRequest {
  message: string;
  character: ChatRequestCharacter;
  episodeId?: string;
  childName?: string;
  childAge?: number;
  /** Structured summary of the child's lesson progress for adaptive AI responses */
  progressSummary?: string;
}